coupla stuff
Began integration of "game" layer.
Set goals for game functions, as well as UI comprehension.

At an given time, if the game is not calculating, the game will be in a specific "waiting" state.
Depending on the waiting state, the next clikc will trigger a UI comprehension FN, and pass its output into the appropriate subsequent fn, which will ultimately resolve at a (possibly) different waiting phase.